+++ Sources/SendBirdSDK.xcframework/ios-arm64_i386_x86_64-simulator/SendBirdSDK.framework/Headers/SBDAppleCriticalAlertOptions.h
@@ -8,6 +8,8 @@
 
 #import <Foundation/Foundation.h>
 
+#import "SBDMappable.h"
+
 NS_ASSUME_NONNULL_BEGIN
 
 /// Represents the options for the Apple Critical Alert.
@@ -15,7 +17,7 @@ NS_ASSUME_NONNULL_BEGIN
 /// When the message that has this options is received by the iOS device, the message is displayed as a critical alert.
 /// See the [`Generating a Remote Notification`](https://developer.apple.com/documentation/usernotifications/setting_up_a_remote_notification_server/generating_a_remote_notification) for the details.
 /// @since 3.0.221
-@interface SBDAppleCriticalAlertOptions : NSObject
+@interface SBDAppleCriticalAlertOptions : NSObject<SBDMappable>
 
 /// The name of a sound file in the iOS app. The default value is `default`.
 /// @since 3.0.221