+++ Sources/SendBirdSDK.xcframework/ios-arm64_i386_x86_64-simulator/SendBirdSDK.framework/Headers/SBDChannelContext.h
@@ -8,23 +8,26 @@
 
 #import <Foundation/Foundation.h>
 
-#import "SBDSource.h"
+#import "SBDCollectionEventSource.h"
 
 NS_ASSUME_NONNULL_BEGIN
 
 /// Represents what happened to the channels.
-/// @since 3.0.227
+/// @since 3.1.0
 @interface SBDChannelContext : NSObject
 
-@property (atomic) SBDSource source;
+/// @since 3.1.0
+@property (atomic) SBDCollectionEventSource source;
 
 /// Initializes with the channel source.
 /// @param source Channel source.
 /// @return SBDChannelContext object.
-- (instancetype)initWithChannelSource:(SBDSource)source;
+/// @since 3.1.0
+- (instancetype)initWithChannelSource:(SBDCollectionEventSource)source;
 
 /// Checks whether the context is from an event or not.
 /// @return If YES, the context is from an event.
+/// @since 3.1.0
 - (BOOL)isFromEvent;
 
 @end