+++ Sources/SendBirdSDK.xcframework/ios-arm64_i386_x86_64-simulator/SendBirdSDK.framework/Headers/SBDMessageCollectionDelegate.h
@@ -19,7 +19,7 @@
 NS_ASSUME_NONNULL_BEGIN
 
 /// Methods that are invoked when the event of the message collection occurs.
-/// @since 3.0.227
+/// @since 3.1.0
 @protocol SBDMessageCollectionDelegate <NSObject>
 
 @optional
@@ -29,6 +29,7 @@ NS_ASSUME_NONNULL_BEGIN
 /// @param context The reason why the messages are added.
 /// @param channel The channel that has the messages.
 /// @param messages Messages to be added.
+/// @since 3.1.0
 - (void)messageCollection:(SBDMessageCollection *)collection
                   context:(SBDMessageContext *)context
                   channel:(SBDGroupChannel *)channel
@@ -39,6 +40,7 @@ NS_ASSUME_NONNULL_BEGIN
 /// @param context The reason why the messages are updated.
 /// @param channel The channel that has the messages.
 /// @param messages Messages to be updated.
+/// @since 3.1.0
 - (void)messageCollection:(SBDMessageCollection *)collection
                   context:(SBDMessageContext *)context
                   channel:(SBDGroupChannel *)channel
@@ -49,30 +51,34 @@ NS_ASSUME_NONNULL_BEGIN
 /// @param context The reason why the messages are deleted.
 /// @param channel The channel that has the messages.
 /// @param messages Messages to be deleted.
+/// @since 3.1.0
 - (void)messageCollection:(SBDMessageCollection *)collection
                   context:(SBDMessageContext *)context
                   channel:(SBDGroupChannel *)channel
           deletedMessages:(NSArray<SBDBaseMessage *> *)messages;
 
-/// Invoked when the channel that has the message collection is changed.
+/// Invoked when the channel that has the message collection is updated.
 /// @param collection Message collection object.
 /// @param context The reason why the channel are changed.
 /// @param channel The channel that has the message collection has.
+/// @since 3.1.0
 - (void)messageCollection:(SBDMessageCollection *)collection
                   context:(SBDMessageContext *)context
-           changedChannel:(SBDGroupChannel *)channel;
+           updatedChannel:(SBDGroupChannel *)channel;
 
 /// Invoked when the channel that has the message collection is deleted.
 /// @param collection Message collection object.
 /// @param context The reason why the channel are deleted.
 /// @param channelUrl Channel URL that has been deleted.
+/// @since 3.1.0
 - (void)messageCollection:(SBDMessageCollection *)collection
                   context:(SBDMessageContext *)context
            deletedChannel:(NSString *)channelUrl;
 
 /// Invoked when the huge gap has been detected.
 /// @param collection Message collection object.
-- (void)didDetectHugeGapInMessageCollection:(SBDMessageCollection *)collection;
+/// @since 3.1.0
+- (void)didDetectHugeGap:(SBDMessageCollection *)collection;
 
 @end